Tweak physics

+++ src/Game.hpp
@@ -30,10 +30,10 @@ private:
   PolygonObject game_character;
   std::vector<PolygonObject> objects;
 
-  uint64_t current_tick = 1000;
-  uint64_t tick_of_jump_request = 0;
-  uint64_t tick_of_last_floor_collision = 0;
-  uint64_t tick_of_last_wall_collision = 0;
+  uint32_t current_tick = 1000;
+  uint32_t tick_of_jump_request = 0;
+  uint32_t tick_of_last_floor_collision = 0;
+  uint32_t tick_of_last_wall_collision = 0;
   bool wall_jump_to_right = false;
 
   glm::vec2 right_direction{0, 0};
